Add Nickname logging (#162)
* Add Nickname logging

This tries to close #161

 - setting.ts added field for log files
 - events/README.md updated link to point to correct page
 - guildMemberUpdate.ts added logic at the end of the event handler
 - setting.ts added command registration for setting log channel for this feature

I currently have the user id as part of the message though that might be overkill.
I am also wondering if I should change the event to avoid a bug case. marked with a
TODO. If this case isn't possible then that part can just be removed and ignored.

* add pr changes