163 bug leave suddenly stopped showing when a person joined occasionally works (#168)
* added Components v2

* Nickname log format change (#165)

* Add Nickname logging

This tries to close #161

 - setting.ts added field for log files
 - events/README.md updated link to point to correct page
 - guildMemberUpdate.ts added logic at the end of the event handler
 - setting.ts added command registration for setting log channel for this feature

I currently have the user id as part of the message though that might be overkill.
I am also wondering if I should change the event to avoid a bug case. marked with a
TODO. If this case isn't possible then that part can just be removed and ignored.

* add pr changes

* feat: Update message format changes

I made the change mimic the style that was requested.

---------

Co-authored-by: William Bowen <william@progress.win>

* file rename

* added mute command (#167)

it sets mute, will log the mute, will send a message to the muted user,
and if the user and muter are in the same voice channel it will send a
message in the chat too.

* updated command response

* readme fix

* updated welcome button

---------

Co-authored-by: Marcus Doucette <36041181+mxd1760@users.noreply.github.com>